pyproject.toml
[project]
name = "unified-mcp"
version = "1.0.0"
description = "Unified dynamic MCP server with progressive discovery for code understanding, documentation, browser automation, and memory"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["mcp", "model-context-protocol", "ai", "code-understanding", "documentation", "browser-automation"]

dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "real_ladybug>=0.13.0",
    "graphiti-core>=0.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-timeout>=2.0.0",
    "pytest-mock>=3.0.0",
    "coverage>=7.0.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["core", "handlers"]
omit = ["tests/*", "examples/*"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
